Add files via upload
This is the login account test code. Because we didn't know about the problems of different versions of PHP before, the test code here will have fatal errors after it is placed on the server. Therefore, according to the latest version of PHP, MySQL is needed here_ Change connect to new mysqli to solve the problem